Update entitlementmanagement-post-accesspackageresourcerequests.md
Added a Note just below the permissions table, to notify people that additional permissions might be required. This was not immediately obvious to me. This note would have saved me quite some time. Only found out this was the case after a chat with Mark Wahl on the Identity Advisors group.